refactor: 상속을 이용한 중복 코드 제거
